requirements.txt
cocos2d==0.6.5
future==0.18.1
numpy==1.17.4
pyglet==1.3.2
six==1.12.0
